requirements.txt
mysqlclient == 1.4.4
Django == 2.2.6
markdown2 == 2.3.8
django-mdeditor==0.1.16